move _human_size to shared location to eliminate circular imports
