+++ ddtrace/internal/utils/__init__.py
@@ -110,3 +110,16 @@ def set_blocked(block_settings: Optional[dict[str, Any]] = None) -> None:
     from ddtrace.internal.constants import STATUS_403_TYPE_AUTO
 
     core.dispatch("asm.set_blocked", (block_settings or STATUS_403_TYPE_AUTO,))
+
+
+def _human_size(nbytes: float) -> str:
+    """Return a human-readable size."""
+    i = 0
+    suffixes = ["B", "KB", "MB", "GB", "TB"]
+    while nbytes >= 1000 and i < len(suffixes) - 1:
+        nbytes /= 1000.0
+        i += 1
+    f = ("%.2f" % nbytes).rstrip("0").rstrip(".")
+    return "%s%s" % (f, suffixes[i])
+
+
