+++ ddtrace/internal/writer/__init__.py
@@ -8,7 +8,6 @@ from .writer import LogWriter
 from .writer import NativeWriter
 from .writer import Response
 from .writer import TraceWriter
-from .writer import _human_size
 from .writer import create_trace_writer
 from .writer_client import WriterClientBase
 
@@ -24,6 +23,5 @@ __all__ = [
     "Response",
     "TraceWriter",
     "WriterClientBase",
-    "_human_size",
     "create_trace_writer",
 ]
